Move reconciliation logic into convergence.go
